requirements.txt
aiohttp>=3.8.0
websockets>=11.0
pyaudio
openai
python-dotenv
numpy
mss
Pillow 